Create peer with custom parameters

+++ src/main/java/com/wireguard/external/wireguard/WgTool.java
@@ -3,20 +3,24 @@ package com.wireguard.external.wireguard;
 import com.wireguard.external.shell.ShellRunner;
 import com.wireguard.external.wireguard.dto.CreatedPeer;
 import com.wireguard.parser.WgShowDumpParser;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
-import java.util.Scanner;
+import java.util.*;
+import java.util.function.Function;
 
 @Profile("prod")
 @Component
 public class WgTool {
+    private static final Logger logger = LoggerFactory.getLogger(ShellRunner.class);
 
     private static final String WG_SHOW_DUMP_COMMAND = "wg show %s dump";
     private static final String WG_GENKEY_COMMAND = "wg genkey";
     private static final String WG_PUBKEY_COMMAND =  "echo %s | wg pubkey";
     private static final String WG_PRESHARED_KEY_COMMAND = "wg genpsk";
-    private static final String WG_ADD_PEER_COMMAND = "wg set %s peer %s preshared-key %s allowed-ips %s persistent-keepalive %d";
+    private static final String WG_ADD_PEER_COMMAND = "wg set %s peer %s";
     private static final String CREATE_FILE_COMMAND = "echo %s > %s";
     private static final String DELETE_FILE_COMMAND = "rm %s";
     private static final String WG_DEL_PEER_COMMAND = "wg set %s peer %s remove";
@@ -28,6 +32,7 @@ public class WgTool {
 
     private String run(String commandStr, Boolean privileged) {
         String[] command = getCommand(commandStr, privileged);
+        logger.debug("Running command: %s".formatted(String.join(" ", command)));
         return shell.execute(command).strip();
     }
     protected String[] getCommand(String commandStr, Boolean privileged) {
@@ -89,20 +94,61 @@ public class WgTool {
 
     //I don't know how to do this without creating a file (wg set doesn't accept preshared key as a parameter)
     public void addPeer(String interfaceName, CreatedPeer peer) {
-        createFile(presharedKeyPath, peer.getPresharedKey());
-        try{
-            run(WG_ADD_PEER_COMMAND.formatted(
+        StringBuilder command = new StringBuilder();
+        command.append(WG_ADD_PEER_COMMAND.formatted(
                 interfaceName,
-                peer.getPublicKey(),
-                presharedKeyPath,
-                String.join(",", peer.getAddress()),
-                peer.getPersistentKeepalive()), true);
+                peer.getPublicKey()));
+        if (!peer.getPresharedKey().isEmpty())
+            createFile(presharedKeyPath, peer.getPresharedKey());
+        List<Argument> arguments = List.of(
+                new Argument("preshared-key",
+                        peer.getPresharedKey().isEmpty() ? null : presharedKeyPath),
+                new Argument("allowed-ips", String.join(",", peer.getAddress()), (value) -> String.join(",", peer.getAddress())),
+                new Argument("persistent-keepalive", String.valueOf(peer.getPersistentKeepalive()))
+        );
+        appendArgumentsIfPresentAndNotEmpty(arguments, command);
+
+        try{
+            run(command.toString(), true);
         } finally {
-            deleteFile(presharedKeyPath);
+            if (!peer.getPresharedKey().isEmpty())
+                deleteFile(presharedKeyPath);
         }
         saveConfig(interfaceName);
     }
 
+    private static class Argument{
+        private final String name;
+        private final String value;
+        private final Function<String, String> valueTransformer;
+
+        public Argument(String name, String value, Function<String, String> valueTransformer) {
+            this.name = name;
+            this.value = value;
+            this.valueTransformer = valueTransformer;
+        }
+
+        public Argument(String name, String value) {
+            this(name, value, (v) -> v);
+        }
+
+        public boolean isPresent(){
+            return value != null && !value.isEmpty();
+        }
+
+        public String getCommand(){
+            return String.format(" %s %s", name, valueTransformer.apply(value));
+        }
+    }
+
+    private void appendArgumentsIfPresentAndNotEmpty(List<Argument> arguments, StringBuilder command) {
+        for (Argument argument: arguments) {
+            if (argument.isPresent()) {
+                command.append(argument.getCommand());
+            }
+        }
+
+    }
     private void saveConfig(String interfaceName) {
         run(WG_SAVE_COMMAND.formatted(interfaceName), true);
     }
